SPL-16009 Move dedup logic to lib/dedup.ts to minimize upstream-file diff
